Move smart folder related code into SmartFolderManager
