+++ map/src/menu/actions/GroupActions.jsx
@@ -11,7 +11,7 @@ import { apiPost } from '../../util/HttpApi';
 import AppContext from '../../context/AppContext';
 import { useTranslation } from 'react-i18next';
 import { SMART_TYPE } from '../share/shareConstants';
-import { populateSmartFolderFiles } from '../../util/hooks/useInitialFilesLoad';
+import { populateSmartFolderFiles } from '../../manager/SmartFoldersManager';
 import { GARMIN_FOLDER_NAME, fetchGarminStatus } from '../../login/garmin/garminApi';
 
 const GroupActions = forwardRef(({ group, setOpenActions, setProcessDownload }, ref) => {
