+++ map/src/menu/tracks/CloudTrackGroup.jsx
@@ -14,7 +14,7 @@ import DividerWithMargin from '../../frame/components/dividers/DividerWithMargin
 import ThreeDotsButton from '../../frame/components/btns/ThreeDotsButton';
 import { fmt } from '../../util/dateFmt';
 import { SMART_TYPE } from '../share/shareConstants';
-import { populateSmartFolderFiles } from '../../util/hooks/useInitialFilesLoad';
+import { populateSmartFolderFiles } from '../../manager/SmartFoldersManager';
 
 export default function CloudTrackGroup({ index, group }) {
     const ctx = useContext(AppContext);
