+++ map/src/menu/tracks/TrackGroupFolder.jsx
@@ -19,7 +19,7 @@ import isEmpty from 'lodash-es/isEmpty';
 import { DEFAULT_SORT_METHOD } from './TracksMenu';
 import Loading from '../errors/Loading';
 import { SMART_TYPE } from '../share/shareConstants';
-import { populateSmartFolderFiles } from '../../util/hooks/useInitialFilesLoad';
+import { populateSmartFolderFiles } from '../../manager/SmartFoldersManager';
 
 export default function TrackGroupFolder({ folder = null, smartf = null }) {
     const ctx = useContext(AppContext);
