+++ map/src/util/hooks/useInitialFilesLoad.js
@@ -3,17 +3,16 @@ import isEmpty from 'lodash-es/isEmpty';
 import { apiGet } from '../HttpApi';
 import TracksManager, {
     getGpxFiles,
-    filterRegularFolders,
     GPX_FILE_EXT,
     INFO_FILE_EXT,
     preparedGpxFile,
     TRACK_VISIBLE_FLAG,
     EMPTY_FILE_NAME,
 } from '../../manager/track/TracksManager';
+import { loadSmartFolders } from '../../manager/SmartFoldersManager';
 import { addOpenedFavoriteGroups } from '../../manager/FavoritesManager';
 import { getShareWithMe } from '../../manager/ShareManager';
 import { FAVOURITES, GPX } from '../../manager/GlobalManager';
-import { SMART_TYPE } from '../../menu/share/shareConstants';
 import { findInfoFile } from '../../manager/track/TrackAppearanceManager';
 import Utils from '../Utils';
 import { INIT_LOGIN_STATE } from '../../manager/LoginManager';
@@ -100,60 +99,6 @@ export async function loadShareFiles(setShareWithMeFiles) {
 }
 
 /**
- * Fetches smart folders from the server and adds them to tracks groups.
- * Clears the smart folders cache before loading.
- */
-export async function loadSmartFolders(setTracksGroups, setSmartFoldersCache) {
-    const res = await getSmartFolders();
-    setSmartFoldersCache({});
-    const smartFolderGroups = (res ?? []).map((smartFolder) => {
-        return {
-            name: smartFolder.name,
-            fullName: smartFolder.name,
-            type: SMART_TYPE,
-            subfolders: [],
-            groupFiles: [],
-            files: [],
-            realSize: smartFolder.userFilePaths?.length ?? 0,
-            lastModifiedMs: smartFolder.creationTime,
-            lastModifiedDate: null,
-            userFilePaths: smartFolder.userFilePaths ?? [],
-        };
-    });
-
-    setTracksGroups((prev) => {
-        const regularFolders = filterRegularFolders(prev);
-        return [...regularFolders, ...smartFolderGroups];
-    });
-}
-
-/**
- * Populates a smart folder with its actual file objects from the file list.
- * Uses caching to avoid redundant processing. Returns the populated folder.
- */
-export function populateSmartFolderFiles(smartFolder, listFiles, smartFoldersCache, setSmartFoldersCache) {
-    const filesArray = [];
-    const cached = smartFoldersCache?.[smartFolder.name];
-    if (cached) {
-        return cached;
-    }
-    (smartFolder.userFilePaths ?? []).forEach((path) => {
-        const file = listFiles?.find((f) => f.name === path);
-        if (file) {
-            filesArray.push({ ...file, smartFolder: true });
-        }
-    });
-    const populated = {
-        ...smartFolder,
-        groupFiles: filesArray,
-        files: filesArray,
-        realSize: filesArray.length,
-    };
-    setSmartFoldersCache((prev) => ({ ...prev, [smartFolder.name]: populated }));
-    return populated;
-}
-
-/**
  * Custom hook that handles initial loading of cloud files when user logs in.
  * Fetches user's tracks, favorites, smart folders, and shared files.
  * Restores previously visible tracks from localStorage.
@@ -230,11 +175,6 @@ export function useInitialFilesLoad({
             }
         }
     }, [loginUser]);
-}
-
-async function getSmartFolders() {
-    const response = await apiGet(`${process.env.REACT_APP_USER_API_SITE}/mapapi/get-smart-folders`, {});
-    return response?.data || null;
 }
 
 function isFileWithDetails(fileName) {
